Repositorio Integer a Long

+++ src/main/java/net/Jaumebalmes/APIG/Repository/AsistenciaRepository.java
@@ -3,7 +3,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import net.Jaumebalmes.APIG.Entities.Asistencia;
 
-public interface AsistenciaRepository extends CrudRepository<Asistencia, Integer> {
+public interface AsistenciaRepository extends CrudRepository<Asistencia, Long> {
 
     public List<Asistencia> findAll();
 }
